update arguments and payload
