+++ packages/twenty-server/src/modules/calendar/calendar-event-import-manager/jobs/calendar-events-import.job.ts
@@ -1,4 +1,4 @@
-import { Logger, Scope } from '@nestjs/common';
+import { Scope } from '@nestjs/common';
 
 import { Processor } from 'src/engine/integrations/message-queue/decorators/processor.decorator';
 import { MessageQueue } from 'src/engine/integrations/message-queue/message-queue.constants';
@@ -25,8 +25,6 @@ export type CalendarEventsImportJobData = {
   scope: Scope.REQUEST,
 })
 export class CalendarEventsImportJob {
-  private readonly logger = new Logger(CalendarEventsImportJob.name);
-
   constructor(
     private readonly googleCalendarSyncService: CalendarEventsImportService,
     @InjectObjectMetadataRepository(ConnectedAccountWorkspaceEntity)
