+++ packages/twenty-server/src/modules/calendar/calendar-event-import-manager/services/calendar-events-import.service.ts
@@ -1,5 +1,4 @@
-import { Injectable, Logger } from '@nestjs/common';
-import { EventEmitter2 } from '@nestjs/event-emitter';
+import { Injectable } from '@nestjs/common';
 
 import { Any } from 'typeorm';
 
@@ -7,46 +6,28 @@ import { BlocklistRepository } from 'src/modules/connected-account/repositories/
 import { ConnectedAccountWorkspaceEntity } from 'src/modules/connected-account/standard-objects/connected-account.workspace-entity';
 import { InjectObjectMetadataRepository } from 'src/engine/object-metadata-repository/object-metadata-repository.decorator';
 import { BlocklistWorkspaceEntity } from 'src/modules/connected-account/standard-objects/blocklist.workspace-entity';
-import { InjectMessageQueue } from 'src/engine/integrations/message-queue/decorators/message-queue.decorator';
-import { MessageQueue } from 'src/engine/integrations/message-queue/message-queue.constants';
-import { MessageQueueService } from 'src/engine/integrations/message-queue/services/message-queue.service';
 import { InjectWorkspaceRepository } from 'src/engine/twenty-orm/decorators/inject-workspace-repository.decorator';
 import { WorkspaceRepository } from 'src/engine/twenty-orm/repository/workspace.repository';
-import { WorkspaceDataSource } from 'src/engine/twenty-orm/datasource/workspace.datasource';
-import { InjectWorkspaceDatasource } from 'src/engine/twenty-orm/decorators/inject-workspace-datasource.decorator';
 import { CalendarEventCleanerService } from 'src/modules/calendar/calendar-event-cleaner/services/calendar-event-cleaner.service';
-import { CalendarEventParticipantService } from 'src/modules/calendar/calendar-event-participant-manager/services/calendar-event-participant.service';
 import { CalendarChannelEventAssociationWorkspaceEntity } from 'src/modules/calendar/common/standard-objects/calendar-channel-event-association.workspace-entity';
 import { CalendarChannelWorkspaceEntity } from 'src/modules/calendar/common/standard-objects/calendar-channel.workspace-entity';
-import { CalendarEventWorkspaceEntity } from 'src/modules/calendar/common/standard-objects/calendar-event.workspace-entity';
 import { CalendarChannelSyncStatusService } from 'src/modules/calendar/common/services/calendar-channel-sync-status.service';
-import { GoogleCalendarGetEventsService } from 'src/modules/calendar/calendar-event-import-manager/drivers/google-calendar/services/google-calendar-get-events.service';
 import { filterEventsAndReturnCancelledEvents } from 'src/modules/calendar/calendar-event-import-manager/utils/filter-events.util';
-import { injectIdsInCalendarEvents } from 'src/modules/calendar/calendar-event-import-manager/utils/inject-ids-in-calendar-events.util';
 import { CalendarSaveEventsService } from 'src/modules/calendar/calendar-event-import-manager/services/calendar-save-events.service';
+import { CalendarGetCalendarEventsService } from 'src/modules/calendar/calendar-event-import-manager/services/calendar-get-events.service';
 
 @Injectable()
 export class CalendarEventsImportService {
-  private readonly logger = new Logger(CalendarEventsImportService.name);
-
   constructor(
-    @InjectWorkspaceRepository(CalendarEventWorkspaceEntity)
-    private readonly calendarEventRepository: WorkspaceRepository<CalendarEventWorkspaceEntity>,
     @InjectWorkspaceRepository(CalendarChannelWorkspaceEntity)
     private readonly calendarChannelRepository: WorkspaceRepository<CalendarChannelWorkspaceEntity>,
     @InjectWorkspaceRepository(CalendarChannelEventAssociationWorkspaceEntity)
     private readonly calendarChannelEventAssociationRepository: WorkspaceRepository<CalendarChannelEventAssociationWorkspaceEntity>,
     @InjectObjectMetadataRepository(BlocklistWorkspaceEntity)
     private readonly blocklistRepository: BlocklistRepository,
-    @InjectWorkspaceDatasource()
-    private readonly workspaceDataSource: WorkspaceDataSource,
     private readonly calendarEventCleanerService: CalendarEventCleanerService,
-    private readonly calendarEventParticipantService: CalendarEventParticipantService,
-    @InjectMessageQueue(MessageQueue.contactCreationQueue)
-    private readonly messageQueueService: MessageQueueService,
-    private readonly eventEmitter: EventEmitter2,
     private readonly calendarChannelSyncStatusService: CalendarChannelSyncStatusService,
-    private readonly googleCalendarGetEventsService: GoogleCalendarGetEventsService,
+    private readonly getCalendarEventsService: CalendarGetCalendarEventsService,
     private readonly calendarSaveEventsService: CalendarSaveEventsService,
   ) {}
 
@@ -59,10 +40,8 @@ export class CalendarEventsImportService {
       connectedAccount.id,
     );
 
-    const calendarChannelId = calendarChannel.id;
-
     const { calendarEvents, nextSyncCursor } =
-      await this.googleCalendarGetEventsService.getCalendarEvents(
+      await this.getCalendarEventsService.getCalendarEvents(
         connectedAccount,
         calendarChannel.syncCursor,
       );
@@ -87,77 +66,17 @@ export class CalendarEventsImportService {
       (event) => event.externalId,
     );
 
-    const existingCalendarEvents = await this.calendarEventRepository.find({
-      where: {
-        iCalUID: Any(filteredEvents.map((event) => event.iCalUID as string)),
-      },
-    });
-
-    const iCalUIDCalendarEventIdMap = new Map(
-      existingCalendarEvents.map((calendarEvent) => [
-        calendarEvent.iCalUID,
-        calendarEvent.id,
-      ]),
-    );
-
-    const calendarEventsWithIds = injectIdsInCalendarEvents(
+    await this.calendarSaveEventsService.saveCalendarEventsAndEnqueueContactCreationJob(
       filteredEvents,
-      iCalUIDCalendarEventIdMap,
-    );
-
-    // TODO: When we will be able to add unicity contraint on iCalUID, we will do a INSERT ON CONFLICT DO UPDATE
-
-    const existingEventsICalUIDs = existingCalendarEvents.map(
-      (calendarEvent) => calendarEvent.iCalUID,
-    );
-
-    const eventsToSave = calendarEventsWithIds.filter(
-      (calendarEvent) =>
-        !existingEventsICalUIDs.includes(calendarEvent.iCalUID),
-    );
-
-    const eventsToUpdate = calendarEventsWithIds.filter((calendarEvent) =>
-      existingEventsICalUIDs.includes(calendarEvent.iCalUID),
-    );
-
-    const existingCalendarChannelEventAssociations =
-      await this.calendarChannelEventAssociationRepository.find({
-        where: {
-          eventExternalId: Any(
-            calendarEventsWithIds.map((calendarEvent) => calendarEvent.id),
-          ),
-          calendarChannel: {
-            id: calendarChannelId,
-          },
-        },
-      });
-
-    const calendarChannelEventAssociationsToSave = calendarEventsWithIds
-      .filter(
-        (calendarEvent) =>
-          !existingCalendarChannelEventAssociations.some(
-            (association) => association.eventExternalId === calendarEvent.id,
-          ),
-      )
-      .map((calendarEvent) => ({
-        calendarEventId: calendarEvent.id,
-        eventExternalId: calendarEvent.externalId,
-        calendarChannelId,
-      }));
-
-    await this.calendarSaveEventsService.saveCalendarEvents(
-      eventsToSave,
-      eventsToUpdate,
-      calendarChannelEventAssociationsToSave,
-      connectedAccount,
       calendarChannel,
+      connectedAccount,
       workspaceId,
     );
 
     await this.calendarChannelEventAssociationRepository.delete({
       eventExternalId: Any(cancelledEventExternalIds),
       calendarChannel: {
-        id: calendarChannelId,
+        id: calendarChannel.id,
       },
     });
